refactor: Management User 도메인 ServiceDTO 분리(#225)
