--- management/src/main/java/com/foodielog/management/user/dto/request/LoginReq.java
@@ -1,15 +1,0 @@
-package com.foodielog.management.user.dto.request;
-
-import com.foodielog.server._core.customValid.valid.ValidPassword;
-import lombok.Getter;
-
-import javax.validation.constraints.Email;
-
-@Getter
-public class LoginReq {
-    @Email
-    private String email;
-
-    @ValidPassword
-    private String password;
-}
