--- management/src/main/java/com/foodielog/management/user/dto/response/ReissueResp.java
@@ -1,17 +1,0 @@
-package com.foodielog.management.user.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Getter;
-
-@Getter
-public class ReissueResp {
-    private final String accessToken;
-
-    @JsonIgnore
-    private final String refreshToken;
-
-    public ReissueResp(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
